requirements.txt
mysql-connector-python
python-decouple
requests
selenium